requirements.txt
Django==4.2.13 #updated
djangorestframework==3.15.2     #updated 
PyJWT==2.8.0
drf-yasg==1.21.7
pycryptodome==3.18.0
pycryptodomex==3.18.0